# [Xamarin.Android.Build.Tasks] Add PerformanceTimer for measuring task steps time.