Running an MSBuild task in a profiler is extremely hard, so for now I've been resorting to "printf" profiling.  This adds a utility class that makes it easier to add performance information to the steps in a task.  The output ends up in a `.binlog`.
 
Example output (MAUI template release build):

```
- BuildApk.RunTask - 13597ms
  - Setup - 0ms
  - Compression - 0ms
  - DSOWrapperGenerator - 0ms
  - ExecuteWithAbi 'arm64-v8a, x86_64' - 13591ms
    - Setup - 24ms
    - Add DalvikClasses - 1933ms
    - Embed Assemblies - 5120ms
      - Add 156 user assemblies - 2420ms
      - Add 80 framework assemblies - 1739ms
      - Generate assembly store - 42ms
      - Add assembly store - 918ms
    - AddRuntimeConfigBlob - 197ms
    - AddRuntimeLibraries - 21ms
    - Flush - 169ms
    - AddNativeLibraries - 829ms
    - AddAdditionalNativeLibraries - 0ms
    - Add files - 42ms
    - Add Jars - 515ms
    - Clean up Removed files - 0ms
    - Flush - 4640ms
    - FixupArchive - 93ms
  - Cleanup - 2ms
```

This is roughly what I used to profile the `GenerateJavaStubs` task (https://github.com/dotnet/android/pull/9208, https://github.com/dotnet/android/pull/9215) and I am now starting to look and see if there are any performance wins in `BuildApk`.  I use this locally, the team can decide if we want to commit this to the repo or not.